tidy_app {formatR}	R Documentation

A Shiny app to format R code

Description

Run a Shiny app that formats R code via tidy_source(). This app uses input widgets, such as checkboxes, to pass arguments to tidy_source().

Usage


tidy_app()

Examples


if (interactive()) formatR::tidy_app()

[Package formatR version 1.14.1 Index]