tidy_eval {formatR}	R Documentation

Evaluate R code and mask the output by a prefix

Description

This function is designed to insert the output of each chunk of R code into the source code without really breaking the source code, since the output is masked in comments.

Usage

tidy_eval(source = "clipboard", ..., file = "", prefix = "## ", envir = parent.frame())

Arguments

source
the input filename (by default the clipboard; see tidy_source)

...
other arguments passed to tidy_source

file
the file to write by cat; by default the output is printed on screen

prefix
the prefix to mask the output

envir
the environment in which to evaluate the code (by default the parent environment; if we do not want to mess up with the parent environment, we can set envir = NULL or envir = new.env())

Value

Evaluated R code with corresponding output (printed on screen or written in a file).

References

https://yihui.name/formatR

Examples

library(formatR)
## evaluate simple code as a character vector
tidy_eval(text = c("a<-1+1;a", "matrix(rnorm(10),5)"))

## evaluate a file
tidy_eval(system.file("format", "messy.R", package = "formatR"))

[Package formatR version 1.5 Index]