tidy_rstudio {formatR}	R Documentation

Reformat R code in RStudio IDE

Description

If any R code is selected in the RStudio source editor, this function reformats the selected code; otherwise it reformats the current open file (if it is unsaved, it will be automatically saved).

Usage


tidy_rstudio(...)

Arguments

...
Arguments to be passed to tidy_source(), among which the indent argument will respect the value you set for the number of spaces for indentation in RStudio.

Note

If the output is not what you want, you can undo the change in the editor (Ctrl + Z or Command + Z).

Examples


formatR::tidy_rstudio()
formatR::tidy_rstudio(args.newline = TRUE)


[Package formatR version 1.14.1 Index]