tidy_dir {formatR}	R Documentation

Format all R scripts under a directory, or specified R scripts

Description

Look for all R scripts under a directory (using the pattern "[.][RrSsQq]$"), then tidy them with tidy_source(). If successful, the original scripts will be overwritten with reformatted ones. Please back up the original directory first if you do not fully understand the tricks used by tidy_source(). tidy_file() formats scripts specified by file names.

Usage


tidy_dir(path = ".", recursive = FALSE, ...)

tidy_file(file, ...)

Arguments

path
The path to a directory containning R scripts.

recursive
Whether to recursively look for R scripts under path.

...
Other arguments to be passed to tidy_source().

file
A vector of filenames.

Value

Invisible NULL.

Author(s)

Yihui Xie (tidy_dir) and Ed Lee (tidy_file)

See Also

tidy_source()

Examples


library(formatR)

path = tempdir()
file.copy(system.file("demo", package = "base"), path, recursive = TRUE)
tidy_dir(path, recursive = TRUE)

[Package formatR version 1.14.1 Index]