usage {formatR}	R Documentation

Show the usage of a function

Description

Print the reformatted usage of a function. The arguments of the function are searched by argsAnywhere, so the function can be either exported or non-exported in a package. S3 methods will be marked.

Usage

usage(FUN, width = getOption("width"), tidy = TRUE, output = TRUE)

Arguments

FUN
the function name

width
the width of output (passed to width.cutoff in tidy_source)

tidy
whether to reformat the usage code

output
whether to write the output to the console (via cat)

Value

The R code for the usage is returned as a character string (invisibly).

See Also

tidy_source

Examples

library(formatR)
usage(var)

usage(plot)

usage(plot.default)  # default method
usage("plot.lm")  # on the 'lm' class

usage(usage)

usage(barplot.default, width = 60)  # narrower output

[Package formatR version 1.4 Index]